feat(routes): support all routes
