+++ app/src/components/Menu.tsx
@@ -7,43 +7,24 @@ import {
   Gear,
   Globe,
   Laptop,
-  PencilSimple,
   Plus,
   Question,
   Share,
   TreeStructure,
   User,
 } from "phosphor-react";
-import { memo, useContext, useRef, useState } from "react";
-import { useForm } from "react-hook-form";
-import { useMutation } from "react-query";
+import { memo, useContext } from "react";
 import { useHistory, useRouteMatch } from "react-router-dom";
 
 import { gaChangeTab, gaCopyChart, gaNewChart } from "../lib/analytics";
-import {
-  isError,
-  randomChartName,
-  slugify,
-  titleToLocalStorageKey,
-} from "../lib/helpers";
-import {
-  useCurrentHostedChart,
-  useIsHelp,
-  useIsReadOnly,
-  useIsValidSponsor,
-  useTitle,
-} from "../lib/hooks";
-import { makeChart, queryClient, renameChart } from "../lib/queries";
+import { randomChartName, titleToLocalStorageKey } from "../lib/helpers";
+import { useIsHelp, useIsReadOnly, useTitle } from "../lib/hooks";
 import { Box, BoxProps, Type } from "../slang";
-import { AppContext, Showing, useSession } from "./AppContext";
+import { AppContext, Showing } from "./AppContext";
 import { ReactComponent as BrandSvg } from "./brand.svg";
 import styles from "./Menu.module.css";
+import { RenameButton } from "./RenameButton";
 import {
-  Button,
-  Dialog,
-  Input,
-  Notice,
-  Section,
   smallBtnTypeSize,
   smallIconSize,
   Tooltip,
@@ -251,7 +232,7 @@ const WorkspaceSection = memo(function WorkspaceSection({
         </Box>
       </Box>
       <Box flow="column" gap={1}>
-        {!isReadOnly && !isHelp && <RenameButton fullText={fullText} />}
+        {/* {!isReadOnly && !isHelp && <RenameButton fullText={fullText} />} */}
         {isReadOnly ? <CloneButton fullText={fullText} /> : <ExportButton />}
       </Box>
     </Box>
@@ -271,140 +252,6 @@ function translatedTitle(current: Showing) {
     default:
       return current;
   }
-}
-
-function RenameButton({ fullText }: { fullText: string }) {
-  const isValidSponsor = useIsValidSponsor();
-  const session = useSession();
-  const [initialName, isHosted] = useTitle();
-  const { data } = useCurrentHostedChart();
-  const [dialog, setDialog] = useState(false);
-  const { push } = useHistory();
-  const { register, handleSubmit, watch, formState } = useForm<{
-    name: string;
-    convertToHosted?: boolean;
-  }>({
-    defaultValues: { name: initialName, convertToHosted: false },
-    mode: "onChange",
-  });
-  const convertToHosted = watch("convertToHosted");
-  const currentName = watch("name");
-  const { ref, ...rest } = register("name", {
-    required: true,
-    minLength: 2,
-    setValueAs: (z) => (isHosted || convertToHosted ? z : slugify(z)),
-  });
-  const inputRef = useRef<null | HTMLInputElement>(null);
-  const rename = useMutation(
-    "updateChartName",
-    async ({
-      name,
-      convertToHosted,
-    }: {
-      name: string;
-      convertToHosted?: boolean;
-    }) => {
-      if (isHosted && data) {
-        await renameChart(data.id, name);
-      } else if (convertToHosted) {
-        if (session?.user?.id) {
-          const response = await makeChart({
-            name,
-            user_id: session?.user?.id,
-            chart: fullText,
-          });
-          if (!response) throw new Error("Could not create hosted chart");
-          const charts = response.data;
-          if (!charts) throw new Error("Could not create hosted chart");
-          const chart = charts[0];
-          if (!chart) throw new Error("Could not create hosted chart");
-          push(`/u/${chart.id}`);
-        }
-      } else {
-        const oldKey = titleToLocalStorageKey(slugify(initialName));
-        const newSlug = slugify(name);
-        const newKey = titleToLocalStorageKey(newSlug);
-        if (window.localStorage.getItem(newKey) !== null)
-          throw new Error("Chart already exists");
-        window.localStorage.setItem(newKey, fullText);
-        push(`/${newSlug}`);
-        window.localStorage.removeItem(oldKey);
-      }
-    },
-    {
-      onSuccess: () => {
-        setDialog(false);
-        if (isHosted) queryClient.resetQueries(["useChart"]);
-      },
-    }
-  );
-  const isValid =
-    formState.isValid && (currentName !== initialName || convertToHosted);
-  return (
-    <>
-      <Tooltip
-        label={t`Rename`}
-        aria-label={t`Rename`}
-        className={`slang-type size-${tooltipSize}`}
-      >
-        <Button
-          style={{ minWidth: 0 }}
-          onClick={() => setDialog(true)}
-          aria-label={t`Rename`}
-        >
-          <PencilSimple size={smallIconSize} />
-        </Button>
-      </Tooltip>
-      <Dialog
-        dialogProps={{
-          isOpen: dialog,
-          onDismiss: () => setDialog(false),
-          initialFocusRef: inputRef,
-          "aria-label": t`Rename`,
-        }}
-        innerBoxProps={{
-          as: "form",
-          onSubmit: handleSubmit((data) => rename.mutate(data)),
-        }}
-      >
-        <Section>
-          <Type as="h2" weight="400">
-            <Trans>Rename</Trans>
-          </Type>
-          {isValidSponsor && !isHosted ? (
-            <Box
-              flow="column"
-              gap={2}
-              content="normal start"
-              items="center normal"
-            >
-              <Type size={-1}>
-                <Trans>Convert to hosted chart?</Trans>
-              </Type>
-              <input type="checkbox" {...register("convertToHosted")} />
-            </Box>
-          ) : null}
-          <Input
-            {...rest}
-            ref={(el) => {
-              ref(el);
-              inputRef.current = el;
-            }}
-            isLoading={rename.isLoading}
-          />
-          <Box flow="column" content="normal space-between">
-            <Button
-              type="button"
-              text={`Cancel`}
-              onClick={() => setDialog(false)}
-            />
-            <Button type="submit" text={t`Submit`} disabled={!isValid} />
-          </Box>
-          {isError(rename.error) && <Notice>{rename.error.message}</Notice>}
-        </Section>
-      </Dialog>
-    </>
-  );
 }
 
 function ExportButton() {
