+++ app/src/slang/config.ts
@@ -35,6 +35,7 @@ export const colors = {
   brandHsl: "233, 100%, 68%",
   darkHighlight: palette.purple[4],
   headerBtnHover: palette.purple[3],
+  greenHsl: "102, 99%, 43%",
 };
 
 export const darkTheme: typeof colors = {
@@ -54,6 +55,7 @@ export const darkTheme: typeof colors = {
   brandHsl: "233, 100%, 68%",
   darkHighlight: palette.purple[2],
   headerBtnHover: palette.purple[3],
+  greenHsl: "102, 99%, 43%",
 };
 
 const config: Partial<SlangConfig> = {
