+++ app/src/slang/index.tsx
@@ -26,6 +26,7 @@ type Colors =
   | "color-brandHsl"
   | "color-darkHighlight"
   | "color-headerBtnHover"
+  | "color-greenHsl"
   | "palette-white-0"
   | "palette-white-1"
   | "palette-white-2"
