+++ app/src/slang/slang.css
@@ -640,6 +640,7 @@ body {
 	--color-brandHsl:233, 100%, 68%;
 	--color-darkHighlight:#001EFF;
 	--color-headerBtnHover:#EFF0FF;
+	--color-greenHsl:102, 99%, 43%;
  }
 
 .slang-box.flow {
